wallet: move master key.
